requirements.txt
flask
flask-cors
python-keycloak
gunicorn
pyjwt
requests